+++ packages/kit/test/types.d.ts
@@ -1,4 +1,4 @@
-/// <reference types="@sveltejs/kit" />
+/// <reference types="../types/index.js" />
 
 import { Page, Response as PlaywrightResponse } from 'playwright-chromium';
 import { RequestInfo, RequestInit, Response as NodeFetchResponse } from 'node-fetch';
@@ -31,7 +31,7 @@ export type TestContext = {
 	};
 
 	watcher: any; // watcher type is not exposed
-	server: import('http').Server;
+	server: import('net').Server;
 	reset: () => Promise<void>;
 	unpatch: () => void;
 };